Refactor atoms handling
