+++ src/quacc/atoms/__init__.py
@@ -1,0 +1,1 @@
+"""Tools for handling and manipulating Atoms objects"""
