+++ src/quacc/atoms/slabs.py
@@ -12,7 +12,7 @@ from pymatgen.analysis.adsorption import AdsorbateSiteFinder
 from pymatgen.core.surface import Slab, center_slab, generate_all_slabs
 from pymatgen.io.ase import AseAtomsAdaptor
 
-from quacc.utils.atoms import copy_atoms
+from quacc.atoms.basics copy_atoms
 
 if TYPE_CHECKING:
     from pymatgen.core import Structure
