+++ src/quacc/recipes/emt/defects.py
@@ -6,9 +6,9 @@ from typing import TYPE_CHECKING
 from pymatgen.analysis.defects.generators import VacancyGenerator
 
 from quacc import flow, job, subflow
+from quacc.atoms.defects import make_defects_from_bulk
 from quacc.recipes.emt.core import relax_job, static_job
 from quacc.schemas import fetch_atoms
-from quacc.utils.defects import make_defects_from_bulk
 
 if TYPE_CHECKING:
     from ase import Atoms
