+++ src/quacc/recipes/emt/slabs.py
@@ -4,9 +4,9 @@ from __future__ import annotations
 from typing import TYPE_CHECKING
 
 from quacc import flow, job, subflow
+from quacc.atoms.slabs import make_slabs_from_bulk
 from quacc.recipes.emt.core import relax_job, static_job
 from quacc.schemas import fetch_atoms
-from quacc.utils.slabs import make_slabs_from_bulk
 
 if TYPE_CHECKING:
     from ase import Atoms
