+++ src/quacc/schemas/atoms.py
@@ -9,7 +9,7 @@ from emmet.core.structure import MoleculeMetadata, StructureMetadata
 from monty.json import jsanitize
 from pymatgen.io.ase import AseAtomsAdaptor
 
-from quacc.utils.atoms import copy_atoms
+from quacc.atoms.basics copy_atoms
 from quacc.utils.dicts import clean_dict
 
 if TYPE_CHECKING:
