+++ src/quacc/utils/__init__.py
@@ -1,6 +1,6 @@
 """
 Various utility functions
 """
-from quacc.utils.atoms import check_charge_and_spin
+from quacc.atoms.basics check_charge_and_spin
 
 __all__ = ["check_charge_and_spin"]
