+++ tests/utils/test_slabs.py
@@ -7,7 +7,7 @@ import pytest
 from ase.build import bulk, fcc100, molecule
 from ase.io import read
 
-from quacc.utils.slabs import (
+from quacc.atoms.slabs import (
     flip_atoms,
     get_surface_energy,
     make_adsorbate_structures,
